tweaks to containers in explore.html and explore.css
Please take a look and let me know of any changes that need to be made. I think I'm about done with the explore html and css and need to take a break from it to work on BIT142 stuff. The footer info is in there and should look correct once the global.css is finished.